generative_ai/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
beanie
motor
pydantic[email]
boto3
dnspython
werkzeug
tzdata 
celery
redis
passlib[bcrypt]
python-jose[cryptography]